# Improve performance of type-checking record literals