@sjakobi 2019-07-02
Nice! I'm wondering whether we should try to do the same with `Union`:

https://github.com/dhall-lang/dhall-haskell/blob/92bdd56a3426bb912aadc0ab9b8a8effe0ec6714/dhall/src/Dhall/TypeCheck.hs#L513-L547

It's a bit more complicated though since `k0` isn't necessary the first key of `kts`. So we'd have to find `k0`'s position and re-insert it at the same position again. Or we could do something terrible like deleting it only from the `Data.Map` but leaving it in the list of keys… :smiling_imp: