@Gabriel439 2019-07-02
@sjakobi: I'm going to do that separately, because I want to make sure that I have an example benchmark that shows that the change is an improvement before making the change.