@sjakobi 2019-07-02
@Gabriel439 No worries. I doubt that it offers similar performance gains as the `RecordLit` case. Should I open an issue to track the idea?